more code reorganization and clean up
