+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/SysResultVisitor.java
@@ -4,7 +4,7 @@ import edu.ucsc.refactor.Issue;
 import edu.ucsc.refactor.NamedLocation;
 import edu.ucsc.refactor.util.Note;
 import edu.ucsc.refactor.cli.results.*;
-import edu.ucsc.refactor.util.Commit;
+import edu.ucsc.refactor.Commit;
 import edu.ucsc.refactor.util.Notes;
 
 import java.util.Iterator;
