+++ ext/vesper-history/src/edu/ucsc/refactor/util/CommitHistory.java
@@ -4,6 +4,7 @@ import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
+import edu.ucsc.refactor.Commit;
 
 import java.util.Iterator;
 import java.util.NavigableSet;
