+++ src/edu/ucsc/refactor/internal/Downstream.java
@@ -1,11 +1,10 @@
 package edu.ucsc.refactor.internal;
 
 import com.google.common.base.Preconditions;
+import edu.ucsc.refactor.Commit;
 import edu.ucsc.refactor.internal.util.AstUtil;
 import edu.ucsc.refactor.spi.CommitSummary;
 import edu.ucsc.refactor.spi.Repository;
-import edu.ucsc.refactor.util.Commit;
-import edu.ucsc.refactor.util.SourceHistory;
 import org.eclipse.jdt.core.dom.ASTNode;
 
 /**
@@ -31,10 +30,6 @@ public class Downstream implements Repository {
         }
 
         return commit;
-    }
-
-    @Override public SourceHistory pull(String historyForId) {
-        return new SourceHistory(); // empty one
     }
 
     @Override public String toString() {
